changed lat lng columns to float

+++ db/migrate/20100428014504_type_change_in_map.rb
@@ -1,7 +1,7 @@
 class TypeChangeInMap < ActiveRecord::Migration
   def self.up
-    change_column :maps, :lat, :float
-    change_column :maps, :lng, :float
+    change_column :maps, :lat, :numeric
+    change_column :maps, :lng, :numeric
   end
 
   def self.down
